# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: 筑享开发团队
  description: 筑享移动端和管理端后端接口。除明确标注为公开接口外，受保护接口需在 Authorization 请求头中携带 Bearer 访问令牌。
  license:
    name: 内部使用
  title: 筑享租房平台接口文档
  version: v1
servers:
- description: Generated server url
  url: http://localhost:18080/api
...
paths:
...
  /houses/{houseId}/favorite:
    delete:
      description: 从当前用户收藏中移除指定房源。
      operationId: unfavorite
      parameters:
      - description: 房源 ID
        example: house_001
        in: path
        name: houseId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFavoriteResult"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 取消收藏房源
      tags:
      - 房源
    post:
      description: 将指定房源加入当前用户收藏。
      operationId: favorite
      parameters:
      - description: 房源 ID
        example: house_001
        in: path
        name: houseId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseFavoriteResult"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 收藏房源
      tags:
      - 房源
...
  /profile:
    get:
      description: 返回当前登录用户的手机号、昵称、头像、角色、实名认证状态和密码设置状态。
      operationId: getProfile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取个人资料
      tags:
      - 个人中心
    put:
      description: 按需更新当前用户昵称和头像 URL；未传字段保持不变。
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 更新个人资料
      tags:
      - 个人中心
  /profile/avatar:
    post:
      description: 上传不超过 5MB 的头像文件，并更新当前用户头像地址。
      operationId: uploadAvatar
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 头像文件，最大 5MB
              required:
              - file
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAvatarResult"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 上传头像
      tags:
      - 个人中心
  /profile/current-home:
    get:
      description: 返回当前生效租约对应的房源、房间和门锁摘要；无生效租约时数据可能为空。
      operationId: currentHome
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCurrentHome"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取当前住所
      tags:
      - 个人中心
  /profile/favorite-houses:
    get:
      description: 分页返回当前用户收藏的房源。
      operationId: favoriteHouses
      parameters:
      - description: 页码，从 1 开始
        example: 1
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int64
          default: 1
          minimum: 1
      - description: 每页条数，范围 1-100
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
          format: int64
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePageDataHouseView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取收藏房源
      tags:
      - 个人中心
  /profile/lock:
    get:
      description: 返回当前生效租约关联的门锁状态和开锁权限有效期。
      operationId: lockInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLockInfo"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 获取当前门锁
      tags:
      - 个人中心
  /profile/password:
    put:
      description: 校验旧密码后设置新密码，新密码长度为 6-32 位。
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 修改密码
      tags:
      - 个人中心
  /profile/password/set:
    put:
      description: 当前账号尚无登录密码时直接设置新密码；已设置密码的账号需使用修改密码接口。
      operationId: setPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 首次设置密码
      tags:
      - 个人中心
  /profile/phone:
    put:
      description: 使用新手机号收到的短信验证码完成换绑。
      operationId: changePhone
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePhoneRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUserView"
          description: 请求成功
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求参数错误或业务校验失败
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 未登录、访问令牌缺失或已失效
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 当前用户无权执行该操作
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 请求的业务资源不存在
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 资源状态冲突，无法完成当前操作
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
          description: 服务器内部错误
      security:
      - bearerAuth: []
      summary: 修改手机号
      tags:
      - 个人中心
...
components:
  schemas:
...
    ApiResponseAvatarResult:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/AvatarResult"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseCurrentHome:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/CurrentHome"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
    ApiResponseFavoriteResult:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/FavoriteResult"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseLockInfo:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/LockInfo"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
    ApiResponsePageDataHouseView:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/PageDataHouseView"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    ApiResponseUserView:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          $ref: "#/components/schemas/UserView"
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
    ApiResponseVoid:
      type: object
      description: 统一接口响应
      properties:
        code:
          type: integer
          format: int32
          description: 业务状态码，与 HTTP 状态码保持一致
          example: 200
        data:
          description: 响应数据；无返回数据时为 null
        message:
          type: string
          description: 处理结果说明
          example: success
...
    AvatarResult:
      type: object
      description: 头像上传结果
      properties:
        avatarUrl:
          type: string
          description: 头像图片 URL
...
    ChangePasswordRequest:
      type: object
      description: 密码修改请求
      properties:
        newPassword:
          type: string
          description: 新密码，长度 6-32 位
          example: NewExample123
          maxLength: 32
          minLength: 6
        oldPassword:
          type: string
          description: 当前密码
          minLength: 1
      required:
      - newPassword
      - oldPassword
    ChangePhoneRequest:
      type: object
      description: 手机号换绑请求
      properties:
        code:
          type: string
          description: 新手机号收到的 6 位验证码
          example: "123456"
          minLength: 1
          pattern: "^\\d{6}$"
        newPhone:
          type: string
          description: 新手机号
          example: "13900139000"
          minLength: 1
          pattern: "^1\\d{10}$"
      required:
      - code
      - newPhone
...
    CurrentHome:
      type: object
      description: 当前住所摘要
      properties:
        building:
          type: string
          description: 楼栋
        community:
          type: string
          description: 小区名称
        houseId:
          type: string
          description: 房源 ID
        leaseId:
          type: string
          description: 租约 ID
        leaseStatus:
          type: string
          description: 租约状态
        lockId:
          type: string
          description: 门锁 ID
        lockStatus:
          type: string
          description: 门锁状态
        room:
          type: string
          description: 房号
        unit:
          type: string
          description: 单元
    FavoriteResult:
      type: object
      description: 房源收藏操作结果
      properties:
        houseId:
          type: string
          description: 房源 ID
        isFavorite:
          type: boolean
          description: 当前用户是否已收藏
...
    HouseView:
      type: object
      description: 房源列表摘要
      properties:
        area:
          type: integer
          format: int32
          description: 建筑面积，单位平方米
        availableDate:
          type: string
          format: date
          description: 最早可入住日期
        community:
          type: string
          description: 小区名称
        coverImage:
          type: string
          description: 封面图片 URL
        decoration:
          type: string
          description: 装修情况
        description:
          type: string
          description: 详细说明
        facilities:
          type: array
          description: 配套设施
          items:
            type: string
        floor:
          type: string
          description: 楼层描述
        id:
          type: string
          description: 记录 ID
        isFavorite:
          type: boolean
          description: 当前用户是否已收藏
        isSmartLockSupported:
          type: boolean
          description: 是否支持智能门锁
        location:
          type: string
          description: 区域或商圈位置
        metro:
          type: string
          description: 附近地铁信息
        orientation:
          type: string
          description: 房屋朝向
        price:
          type: integer
          format: int32
          description: 月租金，单位元
        roomType:
          type: string
          description: 户型
        status:
          type: string
          description: 业务状态
        tags:
          type: array
          description: 房源标签
          items:
            type: string
        title:
          type: string
          description: 标题
...
    LockInfo:
      type: object
      description: 租约关联门锁和权限摘要
      properties:
        batteryLevel:
          type: integer
          format: int32
          description: 门锁电量百分比
        endDate:
          type: string
          description: 结束日期
        leaseId:
          type: string
          description: 租约 ID
        leaseStatus:
          type: string
          description: 租约状态
        lockBrand:
          type: string
          description: 门锁品牌
        lockId:
          type: string
          description: 门锁 ID
        lockName:
          type: string
          description: 门锁名称
        lockStatus:
          type: string
          description: 门锁状态
        permissionStatus:
          type: string
          description: 开锁权限状态
        startDate:
          type: string
          description: 开始日期
        validFrom:
          type: string
          description: 权限生效时间
        validTo:
          type: string
          description: 权限失效时间
...
    PageDataHouseView:
      type: object
      description: 分页数据
      properties:
        hasMore:
          type: boolean
          description: 是否还有下一页
          example: true
        items:
          type: array
          description: 当前页数据
          items:
            $ref: "#/components/schemas/HouseView"
        page:
          type: integer
          format: int64
          description: 当前页码，从 1 开始
          example: 1
        pageSize:
          type: integer
          format: int64
          description: 每页条数
          example: 20
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 42
...
    SetPasswordRequest:
      type: object
      description: 首次设置登录密码请求
      properties:
        newPassword:
          type: string
          description: 首次设置的登录密码，长度 6-32 位
          example: Example123
          maxLength: 32
          minLength: 6
      required:
      - newPassword
...
    UpdateProfileRequest:
      type: object
      description: 个人资料更新请求，字段均可选
      properties:
        avatarUrl:
          type: string
          description: 新头像 URL
          example: /api/uploads/avatar.jpg
          maxLength: 500
          minLength: 0
        nickname:
          type: string
          description: 新昵称，长度 1-30 位
          example: 小筑
          maxLength: 30
          minLength: 1
    UserView:
      type: object
      description: 用户公开及账户资料
      properties:
        avatarUrl:
          type: string
          description: 头像图片 URL
        hasPassword:
          type: boolean
          description: 当前用户是否已设置登录密码
        id:
          type: string
          description: 记录 ID
        isVerified:
          type: boolean
          description: 是否已认证
        nickname:
          type: string
          description: 用户昵称
        phone:
          type: string
          description: 手机号
        role:
          type: string
          description: 用户角色
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "访问令牌，格式：Bearer {accessToken}"
      name: Authorization
      scheme: bearer
      type: http